requirements.txt
selenium==4.10.0
openai==0.27.4
google-api-python-client==2.92.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.28.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0